+++ src/test/MobileSuit/Core/SuitBuildUtils.py
@@ -35,9 +35,45 @@ class MyTestCase(unittest.TestCase):
 
     def test_CreateConverterFactory_with_list_type(self):
         list_str_converter = CreateConverterFactory(List[str], None)(self.mock_context)
-        self.mock_parsing_service.Get.return_value = "parsed_string"
         result = list_str_converter("parsed_string")
         self.assertEqual(result, "parsed_string")
+
+    def test_GetArg_with_string_type(self):
+        def func(arg: str):
+            pass
+        result, step = GetArg(self.mock_parameter, func, "arg", self.mock_context)
+        self.assertEqual(result, "arg")
+        self.assertEqual(step, 1)
+
+    def test_GetArrayArg_with_string_list(self):
+        self.mock_parameter.annotation = list[str]
+        arg_list = ["arg1", "arg2"]
+        result, length = GetArrayArg(self.mock_parameter, self.mock_function, arg_list, self.mock_context)
+        self.assertEqual(result, arg_list)
+        self.assertEqual(length, len(arg_list))
+
+    def test_GetMethodParameterInfo_no_parameters(self):
+        self.mock_function.__annotations__ = {}
+        info = GetMethodParameterInfo(self.mock_function)
+        self.assertEqual(info.TailParameterType, TailParameterType.NoParameter)
+
+    def test_CreateInstance_with_mock_service(self):
+        self.mock_context.GetService.return_value = "mock_service"
+        instance = CreateInstance(str, self.mock_context)
+        self.assertEqual(instance, "mock_service")
+
+    def test_CreateInstanceWithProvider(self):
+        mock_provider = Mock()
+        mock_provider.GetService.return_value = None
+        mock_type = Mock()
+        instance = CreateInstanceWithProvider(mock_type, mock_provider)
+        self.assertIsNone(instance)
+
+    def test_GetArgs_no_arguments(self):
+        self.mock_function.__annotations__ = {}
+        result = GetArgs(self.mock_function, [], self.mock_context)
+        self.assertIsNotNone(result)
+        self.assertEqual(len(result), 0)
 
     def test_get_arg_with_none_arg(self):
         def func(arg: int):
